week-1/postgres-flask-app/requirements.txt
flask
psycopg2-binary
python-dotenv
